test(core): migrate UpdateGroupTests (Wave C)
39 tests, 131 failure sites preserved 1:1. No `throw TestFailure` sites in the legacy file -- straight #expect conversion.
